### 2024-03-09 #6455 @editorialbot
[PRE REVIEW]: nimCSO: A Nim package for Compositional Space Optimization
**Submitting author:** @amkrajewski (<a href="http://orcid.org/0000-0002-2266-0099">Adam Mikolaj Krajewski</a>)
**Repository:** https://github.com/amkrajewski/nimCSO
**Branch with paper.md** (empty if default branch):
**Version:** v0.6.4
