+++ paper/paper.bib
@@ -90,7 +90,7 @@
    The ternary alloys exhibit a single BCC phase which is consistent with CALPHAD calculations as well as Scheil simulations. The present interactive design loop between the ML surrogate model and experiments is demonstrated through the accurate hardness 
    prediction, resulting in cGAN models capable of rapid exploration of the higher-order design space. The hardness of the Mo-Nb-W alloys is in the range of 5-6 GPa due to their solid solution strengthening.},
    author = {Wenjie Li and Lavanya Raman and Arindam Debnath and Marcia Ahn and Shuang Lin and Adam M Krajewski and Shunli Shang and Shashank Priya and Wesley F Reinhart and Zi-Kui Liu and Allison M Beese},
-   doi = {https://dx.doi.org/10.2139/ssrn.4689687},
+   doi = {10.2139/ssrn.4689687},
    keywords = {CALPHAD,Hardness,Machine learning,Refractory multicomponent alloys,Solidification simulation},
    month = {1},
    title = {Design and Validation of Refractory Alloys Using Machine Learning, CALPHAD, and Experiments},
@@ -126,6 +126,7 @@
    month = {2},
    title = {Efficient Generation of Grids and Traversal Graphs in Compositional Spaces towards Exploration and Path Planning Exemplified in Materials},
    url = {http://arxiv.org/abs/2402.03528},
+   doi = {10.48550/arXiv.2402.03528},
    year = {2024},
 }
 
@@ -134,6 +135,7 @@ author = {Goldberg, David E.},
 title = {Genetic Algorithms in Search, Optimization and Machine Learning}, 
 year = {1989}, 
 isbn = {0201157675}, 
+doi = {10.5860/choice.27-0936},
 publisher = {Addison-Wesley Longman Publishing Co., Inc.}, 
 address = {USA}, 
 edition = {1st}, 
