### 2024-03-11 #6455 @amkrajewski
@arfon - I understand, of course. Thanks for the clear and prompt information!

The word count of the manuscript body is **1373** which aligns with several similar recently published JOSS papers. Of course, I can reduce it if necessary. I think the automated count got especially inflated by markdown-formatted tables.
