+++ paper/paper.bib
@@ -84,17 +84,17 @@
 }
 
 @article{Li2024,
-   abstract = {Refractory multicomponent alloys (RMCAs) have garnered attention as potential materials for high-temperature structural applications, due to their excellent mechanical properties. However, conventional alloy design has limitations in terms 
-   of constrained compositional space and a lack of computational databases with adequate coverage. To address this, we present a design framework that leverages machine learning (ML), the CALculation of PHAse Diagram (CALPHAD) method, and experimental 
-   validation to efficiently develop refractory alloys. The present study focuses on the Mo-Nb-W ternary system. Six ternary alloys were inversely designed by means of the conditional generative adversarial network (cGAN) and fabricated via arc melting. 
-   The ternary alloys exhibit a single BCC phase which is consistent with CALPHAD calculations as well as Scheil simulations. The present interactive design loop between the ML surrogate model and experiments is demonstrated through the accurate hardness 
-   prediction, resulting in cGAN models capable of rapid exploration of the higher-order design space. The hardness of the Mo-Nb-W alloys is in the range of 5-6 GPa due to their solid solution strengthening.},
-   author = {Wenjie Li and Lavanya Raman and Arindam Debnath and Marcia Ahn and Shuang Lin and Adam M Krajewski and Shunli Shang and Shashank Priya and Wesley F Reinhart and Zi-Kui Liu and Allison M Beese},
-   doi = {10.2139/ssrn.4689687},
-   keywords = {CALPHAD,Hardness,Machine learning,Refractory multicomponent alloys,Solidification simulation},
-   month = {1},
-   title = {Design and Validation of Refractory Alloys Using Machine Learning, CALPHAD, and Experiments},
-   url = {https://ssrn.com/abstract=4689687},
+   abstract = {Refractory multicomponent alloys (RMCAs) have garnered attention as potential materials for high-temperature structural applications, due to their excellent mechanical properties. However, conventional alloy design has limitations in terms of constrained compositional space and a lack of computational databases with adequate coverage. To address this, we present a design framework that leverages machine learning (ML), the CALculation of PHAse Diagram (CALPHAD) method, and experimental validation to efficiently develop refractory alloys. The present study focuses on the Mo-Nb-W ternary system. Six ternary alloys were inversely designed by means of the conditional generative adversarial network (cGAN) and fabricated via arc melting. The ternary alloys exhibit a single BCC phase which is consistent with CALPHAD calculations as well as Scheil simulations. The present interactive design loop between the ML surrogate model and experiments is demonstrated through the accurate hardness prediction, resulting in cGAN models capable of rapid exploration of the higher-order design space. The hardness of the Mo-Nb-W alloys is in the range of 5-6 GPa due to their solid solution strengthening.},
+   author = {Wenjie Li and Lavanya Raman and Arindam Debnath and Marcia Ahn and Shuang Lin and Adam M. Krajewski and Shunli Shang and Shashank Priya and Wesley F. Reinhart and Zi-Kui Liu and Allison M. Beese},
+   doi = {10.1016/j.ijrmhm.2024.106673},
+   issn = {02634368},
+   journal = {International Journal of Refractory Metals and Hard Materials},
+   keywords = {Machine learning CALPHAD,Refractory multicomponent alloys,Solidification simulation},
+   month = {6},
+   pages = {106673},
+   title = {Design and validation of refractory alloys using machine learning, CALPHAD, and experiments},
+   volume = {121},
+   url = {https://linkinghub.elsevier.com/retrieve/pii/S0263436824001215},
    year = {2024},
 }
 
