### 2024-04-07 #6455 @arfon
@amkrajewski - apologies for not doing this earlier but I think this submission should go for a scope review with the broader editorial team as it's on the smaller-side of what we typically publish at JOSS.

This will likely take a week or two.
