### 2024-04-07 #6455 @amkrajewski
@arfon No problem! Please consider that (a) substantial effort was put into several iterations of optimizing the data structures and algorithms, as evidenced by commit history, and (2) the use of very concise Nim language plus metaprogramming made the codebase easily 2/3 times smaller in terms of LOCs compared to C++/Python.

Let me know if there are any questions!
