### 2024-04-18 #6455 @RMeli
@amkrajewski do you have suggestions for potential reviewers?

Unfortunately `nim` is not listed in the [potential reviewers list](https://reviewers.joss.theoj.org/reviewers), so we currently have no information on which volunteers know the language.
