### 2024-04-19 #6455 @amkrajewski
@arfon Thanks! Great to hear that!

@RMeli Sure! There is an active sub-forum specific to scientific computing using `nim`. I should be able to find volunteers for JOSS there and report back.

In the meantime, based on internal reviews from my colleagues, I believe the codebase should be approachable for anyone familiar with natural sciences, fundamentals of low-level programming, and Python, since `nim`'s syntax is very similar to recent Python versions (including type system as of 3.12). Looking at the JOSS reviewers list and their recent activities, here are some people who may be interested:
- [Henrium](https://github.com/Henrium)
- [atzberg](https://github.com/atzberg)
- [csadorf](https://github.com/csadorf)
- [lys-devel](https://github.com/lys-devel)
- [zeroset](https://github.com/zeroset)
- [bdice](https://github.com/bdice)
- [yangbai90](https://github.com/yangbai90)

I believe [ml-evs](https://github.com/ml-evs) may also be interested and would be a great person to review this, especially in the context of `nimCSO`'s target applications, but we are co-authors on a recent paper representing a possible COI. Although, there are almost 60 authors on it ([arXiv:2402.00572](https://arxiv.org/abs/2402.00572)) so perhaps that COI could be waived under [JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy).
