### 2024-04-19 #6455 @atzberg
I would be happy to try to review.  However, I should mention I am
currently a bit over-subscribed on a few things and I would be available at
the earliest June 3rd, if that works.  -- Paul

On Thu, Apr 18, 2024 at 11:36 PM Rocco Meli ***@***.***>
wrote:

> Thanks @amkrajewski <https://github.com/amkrajewski> for the suggestions.
>
> but we are co-authors on a recent paper representing a possible COI.
> Although, there are almost 60 authors on it (arXiv:2402.00572
> <https://arxiv.org/abs/2402.00572>) so perhaps that COI could be waived
> under JOSS policy
> <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy>
> .
>
> I would tend to agree that co-autorship on a 60 authors paper might be
> waived as a COI. I see you work on different institutions. Would you mind
> expanding a bit more about your interactions with them during the project?
> In any case, let's see if we have other interested reviewers first.
>
