### 2024-05-08 #6731 @Henrium
I have tested in both GitHub Codespaces and Linux, the package is easy to install and works as claimed. I have a few minor comments regarding the remaining checklist items:

- **Summary:** I suggest the following to make it more accessible to "diverse, non-specialist audience": (1) introduce the background first, then what nimCSO is and what it does; (2) elaborate on the purpose and challenges.

- **State of field:** What are some other approaches to compositional space optimization; are there relevant software? References should be added if applicable. It's not necessary to compare with them, but good to make the paper informative.

- In `quickstart.ipynb`: the routine `mostCommon` is clear at first, but got confusing when it comes to "removing elements". What's the optimization objective of removing elements?

- The "Algorithm-Based Search" method relies on an assumption, "elements present in already expanded ...", is it supported by any rationale, experiments, prior studies, etc.?

- I didn't find "community guidelines", though it doesn't seem necessary here. Consider adding one?
