### 2024-05-10 #6731 @amkrajewski
@Henrium Thank you for starting the review quickly and providing great feedback! I appreciate that.

I've opened an issue in the code/paper repository (https://github.com/amkrajewski/nimCSO/issues/2), where I will address your comments one by one.
