### 2024-06-21 dca8193
Fix typos in README.
