### 2024-06-21 #6731 @bdice
Thanks for the ping!

Status update: I started a review above. For any boxes I have not checked above, I have comments in this issue https://github.com/amkrajewski/nimCSO/issues/3. For some comments, I am proposing changes in this PR: https://github.com/amkrajewski/nimCSO/pull/4

I will return to this later -- my review is about 60% complete at the moment.
