### 2024-06-22 amkrajewski/nimCSO#3 @amkrajewski
Hi @bdice Thank you for taking the time to provide such detailed feedback! And, looking at PR #4 , for providing some solutions!

I'm traveling this week for a conference in Italy, so I'm a bit time-limited, but I should be able to revise everything point-by-point over the next two weeks.
