### 2024-06-22 amkrajewski/nimCSO#3 @amkrajewski
> The GitHub Codespaces took a long time to open (>5 minutes). I would try consolidating the conda install commands. The nim packages on conda-forge depend on conda packages of gcc so it is possible to avoid the 'apt install gcc' steps entirely.

Thanks for catching that! I accidentally limited the pre-building scope to the Eastern US. Now, it should start in 30ish seconds worldwide.
