### 2024-06-22 amkrajewski/nimCSO#3 @amkrajewski
> apt-get install nim did not work on my Ubuntu system. Perhaps the "universe" channel is not enabled by default? I see the package here: https://packages.ubuntu.com/noble/nim but it is version 1.6, while this software claims to require Nim 2.0 in nimcso.nimble. Please verify that sufficiently recent versions of nim are available from the recommended installation paths.

Indeed, we require `2.0` in the nimble file, but that's more about what we provide support for (and recommend performance-wise). It should compile fine under `1.6.x` and work as expected. I will add older versions to the text matrix and adjust the install instructions appropriately.
