### 2024-06-23 amkrajewski/nimCSO#3 @amkrajewski
> I looked at the "native Python" benchmark first. It appears that the NumPy benchmark is not measuring the same amount of work as the native Python benchmark. For example, the construction of elMatrix from the input data should be a part of the timing for the NumPy benchmark (and similarly for the nimCSO benchmark). It's okay to compare algorithm runtime that excludes the CSV parsing (such as the cost to create elementalList in the native Python benchmark) but it is not a fair comparison to pre-process that parsed data by converting into a matrix for NumPy or bit-packed array. Make sure that all benchmarks are measuring as close to the same thing as possible when making relative claims about performance. I understand it is important to amortize that cost of creating a bit-packed array for the input data when computing across many element lists, but it's a bit harder to compare to the native Python solution. I suggest trying to find a middle ground, or at least giving more description of what is being measured.

You are right about the pre-processing being excluded and that it *does* skew the benchmark result, but I believe that choice was sound in terms of method comparison.

The benchmark script is just a quick run over 10,000 checks, i.e., orders of magnitude lower number compared to realistic problems in which `elMatrix` calculation time would be negligible. Even for the 10,000 checks, adding  `elMatrix` calculation to the `pythonNumpPy.py` benchmark (extra `_ = np.array([elList2vec(comp) for comp in elementalList], dtype=np.int8)`) costs on the order of 1% (CPU Time [per comparison] 19.2 ns vs 19.4ns)

- [ ] I will do my best to clearly state that within the documentation and manuscript.
