### 2024-06-23 amkrajewski/nimCSO#3 @amkrajewski
> The statement of need describes applications in economics and medicine as well as materials. I understand the materials science application (which is closer to my field of expertise, and is substantiated by the paper) but the economics and medicine use cases need citations.

Thank you for the suggestion!
- [ ] I will include citations for compositional problems in these two areas. Perhaps I will also add some more.
