### 2024-06-23 amkrajewski/nimCSO#3 @amkrajewski
> I'm new to nim and had some confusion while installing. It seems like it is best practice to use nimble install --depsOnly rather than listing the dependencies directly in the README. Also, I tried nimble install and nimble develop but was unsure how to get the nimCSO executable onto my PATH. The compile command nim c -r -f -d:release src/nimcso produces an executable in src/nimCSO.out but that doesn't seem like the proper "final destination." I am thinking of tools like pip install that can place entry points or executables directly onto your PATH.

In general, `nimble` will give you a similar experience to `pip` and allow you to overlay `nim` tools around `nimCSO` without any "manual labor" since `nimCSO` is indexed by it (`PyPI` equivalent) and lets you just install without any cloning or manual compiling.

`nimCSO` (and installation `README`) undergoes such gymnastics and places the binary in the local `src` rather than some more general location because it is meant to be compiled on a per-task basis, taking advantage of optimizations enabled by knowing task specifics. Thus, having it installed would (I think) confuse users and leave unnecessary files behind after task completion.

Thanks for suggesting `nimble install --depsOnly`. When there are only 1-3 dependencies, I generally prefer to clearly state them in the installation instructions and explain what they are, but I can see the appeal of just using `--depsOnly`. I will add them as an alternative to the `README`.

- [ ] I will do my best to clarify the `README` install instructions
- [ ] Add `nimble install --depsOnly` option to `README`
