### 2024-06-24 amkrajewski/nimCSO#3 @amkrajewski
> The repository would benefit from a CONTRIBUTING.md. It does not appear that the repository has had many issues or pull requests. It would be good to explicitly instruct users to file issues and pull requests.

Thank you for the suggestion!

- [ ] I will create contribution guidelines and place them in `CONTRIBUTING.md`

> I strongly recommend configuring a tool like [pre-commit](https://pre-commit.com/) to automatically trim trailing whitespace and format the code. Having enforceable style conventions, even a minimal set, smooths the process of contributing to an open-source project. If you'd like help here, I can explain more and file a PR.

Thank you for the suggestion! I will add a few hooks that should be helpful here (and do the same for some other projects!).

The style conventions in `nim` language are a much more tricky subject compared to most languages. It is *explicitly designed to* not have code style conventions, even on the basic level like naming, and make programmers read code with how it will be parsed into [AST](https://en.wikipedia.org/wiki/Abstract_syntax_tree) in mind. The result is that collaborative projects use camelCase in one file to define a function and then kebab-case in another one to call it. There are some [guidelines for the compiler contributions](https://nim-lang.org/docs/nep1.html) I can add as suggestions, but even the largest projects like [Arraymancer](https://github.com/mratsim/Arraymancer) drink the kool-aid.

- [ ] Add applicable pre-commit hooks like `check-yaml`, `check-case-conflict`, `check-illegal-windows-names`, `end-of-file-fixer`, and `trailing-whitespace`
- [ ] Add `nim` [guidelines for the compiler contributions](https://nim-lang.org/docs/nep1.html) as a suggestion to `CONTRIBUTING.md`
