### 2024-06-24 amkrajewski/nimCSO#3 @amkrajewski
> I think the paper would benefit from a toy example, explaining what it does. It took me some time to understand precisely what is meant by "selecting components" and phrases like "least preventing". Perhaps a toy example could look something like this?

You are 100% right. I was probably too focused on my problem and should have made it broader.

- [ ] I will come up with a general-audience example to demonstrate the functionality of `nimCSO` and problems associated with interdependency.
