### 2024-06-27 amkrajewski/nimCSO#3 @bdice
@amkrajewski Thanks for the preliminary feedback -- I finished my review. I marked where the second half of my review begins above. Overall the paper is in pretty good shape and it should be straightforward to get this to a state that I can approve.

All your responses above sound good.

Also, I marked my PR #4 as ready to review. Ping me when you've had a chance to look at my PR #4 and apply the suggestions above, and I will work on a second round of review.
