### 2024-06-27 #6731 @atzberg
I've started going through the package.  I anticipate by the end of next
week.  -- Paul

On Fri, Jun 21, 2024 at 12:55 PM Bradley Dice ***@***.***>
wrote:

> Thanks for the ping!
>
> Status update: I started a review above. For any boxes I have not checked
> above, I have comments in this issue amkrajewski/nimCSO#3
> <https://github.com/amkrajewski/nimCSO/issues/3>. For some comments, I am
> proposing changes in this PR: amkrajewski/nimCSO#4
> <https://github.com/amkrajewski/nimCSO/pull/4>
>
> I will return to this later -- my review is about 60% complete at the
> moment.
>
