### 2024-07-06 e02ac6f
(GitHub Actions) update `nim` setup action to `v2`
