+++ .github/workflows/documentation.yaml
@@ -23,7 +23,7 @@ jobs:
     steps:
       - name: Checkout
         uses: actions/checkout@v4
-      - uses: jiro4989/setup-nim-action@v1
+      - uses: jiro4989/setup-nim-action@v2
         with:
           nim-version: '2.0.2'
           repo-token: ${{ secrets.GITHUB_TOKEN }}
