+++ .github/workflows/testingOnPush_Windows.yaml
@@ -16,7 +16,7 @@ jobs:
         uses: actions/checkout@v4
 
       - name: Install Nim
-        uses: jiro4989/setup-nim-action@v1
+        uses: jiro4989/setup-nim-action@v2
         with:
           nim-version: '2.0.2'
           repo-token: ${{ secrets.GITHUB_TOKEN }}
