### 2024-07-16 amkrajewski/nimCSO#2 @RMeli
https://github.com/openjournals/joss-reviews/issues/6731
