### 2024-07-22 #6731 @RMeli
Quick update: the author is/was travelling (https://github.com/amkrajewski/nimCSO/pull/4).
