### 2024-07-27 39a9bc4
(tests) added a test matrix with `1.4.x`, `1.6.x`, `2.0.0`, `2.0.x`, and `stable` to the Linux runner tests to verify `nimCSO` runs as expected on legacy versions installed by package managers on some systems, per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2183959682
