### 2024-07-27 55410ee
(tests) small adjustment to go around testing with older unsupported `nim` versions that `nimble install` would complain about
