### 2024-07-27 591088a
(tests) further small adjustment to go around testing with older unsupported `nim` versions
