### 2024-07-27 afea316
(tests) added '1.4.0' to testing with older unsupported `nim` versions
