### 2024-07-27 98c021b
Revert "(tests) added '1.4.0' to testing with older unsupported `nim` versions"
This reverts commit afea316c2e004d6741053cf21e9614bd1a1dd15a.
