+++ .github/workflows/testingOnPush_Linux.yaml
@@ -39,7 +39,6 @@ jobs:
     strategy:
       matrix:
         nim:
-          - '1.4.0'
           - '1.6.x'
           - '2.0.0'
           - '2.0.x'
