### 2024-07-27 85ab1b0
(docs) update installation instructions per JOSS Review Comments #3 and JOSS Review Comments #5
