+++ examples/README.md
@@ -4,12 +4,18 @@ This is a quick start guide to get you up and running with the nimCSO software. 
 
 Alternatively, if you want to start playing with the CLI directly, you can compile the tool using the command below. It will use the default `config.yaml` file pointing to the `dataList.txt` and should take just a few seconds.
 
-    nim c -r -f -d:release src/nimcso
+```sh
+nim c -r -f -d:release src/nimcso
+```
 
 Once the above is run, you will see the `help` printout showing you your options. You can either add them on top of the previous command, which will recompile the tool, like
 
-    nim c -r -f -d:release src/nimcso --geneticSearch
+```sh
+nim c -r -f -d:release src/nimcso --geneticSearch
+```
 
 or use the already compiled runtime
 
-    ./src/nimcso.out -gs
+```sh
+./src/nimcso.out -gs
+```
