### 2024-07-27 991b2dd
(docs) added some clarifying install notes per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2184915606
