### 2024-07-27 11ef4b1
(docs) added `Contributing` section per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2185872399
