### 2024-07-29 cc107dd
(paper) updated benchmarking results after optimization of Python code in PR #4
