+++ paper/paper.md
@@ -40,7 +40,7 @@ bibliography: paper.bib
 
 `nimCSO` is a high-performance tool implementing several methods for selecting components (data dimensions) in compositional datasets, which optimize the data availability and density for applications such as machine learning. Making said choice is a combinatorically hard problem for complex compositions existing in high-dimensional spaces due to the interdependency of components being present. Such spaces are encountered, for instance, in materials science, where datasets on Compositionally Complex Materials (CCMs) often span 20-45 chemical elements, 5-10 processing types, and several temperature regimes, for up to 60 total data dimensions.
 
-At its core, `nimCSO` leverages the metaprogramming ability of the Nim language [@Rumpf2023] to optimize itself at compile time, both in terms of speed and memory handling, to the specific problem statement and dataset at hand based on a human-readable configuration file. As demonstrated in the [Methods and Performance](#methods-and-performance) section, `nimCSO` reaches the physical limits of the hardware (L1 cache latency) and can outperform an efficient native Python implementation over 400 times in terms of speed and 50 times in terms of memory usage (*not* counting interpreter), while also outperforming NumPy implementation 35 and 17 times, respectively, when checking a candidate solution.
+At its core, `nimCSO` leverages the metaprogramming ability of the Nim language [@Rumpf2023] to optimize itself at compile time, both in terms of speed and memory handling, to the specific problem statement and dataset at hand based on a human-readable configuration file. As demonstrated in the [Methods and Performance](#methods-and-performance) section, `nimCSO` reaches the physical limits of the hardware (L1 cache latency) and can outperform an efficient native Python implementation over 100 times in terms of speed and 50 times in terms of memory usage (*not* counting interpreter), while also outperforming NumPy implementation 37 and 17 times, respectively, when checking a candidate solution.
 
 `nimCSO` is designed to be both (1) a user-ready tool, implementing two efficient brute-force approaches (for handling up to 25 dimensions), a custom search algorithm (for up to 40 dimensions), and a genetic algorithm (for any dimensionality), and (2) a scaffold for building even more elaborate methods in the future, including heuristics going beyond data availability. All configuration is done with a simple human-readable `YAML` config file and plain text data files, making it easy to modify the search method and its parameters with no knowledge of programming and only basic command line skills.
 
@@ -67,13 +67,13 @@ Internally, `nimCSO` is built around storing the data and solutions in one of tw
 +----------------+----------------+------------------+-----------------------------+----------------------------+
 | Tool           | Object         | Time per Dataset | Time per Entry *(Relative)* | Database Size *(Relative)* |
 +:===============+:===============+:================:+:===========================:+:==========================:+
-| `Python`^3.11^ | `set`          | 327.4 µs         | 152.3 ns *(x1)*             | 871.5 kB *(x1)*            |
+| `Python`^3.11^ | `set`          | 107.5 µs         | 50.0 ns *(x1)*              | 871.5 kB *(x1)*            |
 +----------------+----------------+------------------+-----------------------------+----------------------------+
-| `NumPy`^1.26^  | `array`        | 40.1 µs          | 18.6 ns  *(x8.3)*           | 79.7 kB  *(x10.9)*         |
+| `NumPy`^1.26^  | `array`        | 36.4 µs          | 16.9 ns  *(x3.0)*           | 79.7 kB  *(x10.9)*         |
[…]
