### 2024-07-29 9075008
(paper) updated benchmarking table description per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2184908310
