### 2024-07-29 51aaab4
(paper) added an acknowledgement for Dr. Dice code contributions
