### 2024-07-29 bde65fc
(misc) added `.pre-commit-config.yaml`
