### 2024-08-07 4127cbb
(docs) added comments to install instructions per #5
