### 2024-08-12 amkrajewski/nimCSO#2 @RMeli
@amkrajewski could you please briefly summarise the progress on the items listed here? Thank you.
