### 2024-08-18 amkrajewski/nimCSO#2 @amkrajewski
- [ ] Improvement of the summary to make it more approachable to the general audience.
- [ ] Elaborate on other methods - reference the Python and NumPy implementations presented within the `benchmarks` directory.
- [ ] Add more references to other scientific domains to broadly present the problem.
- [ ] Clarify "removal of elements" in the quick start guide.
- [ ] Elaborate on the rationale for algorithm assumption.
- [ ] Add "community guidelines" on how to contribute and what.
