### 2024-08-19 5e6c5b8
(paper) added financial modeling application references per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2184909023
