+++ paper/paper.bib
@@ -176,3 +176,35 @@ programming mainstream.}
    title = {Arraymancer v0.7.24: A fast, ergonomic and portable tensor library in Nim with a deep learning focus for CPU, GPU and embedded devices via OpenMP, Cuda and OpenCL backends},
    url = {https://github.com/mratsim/Arraymancer},
 }
+
+@article{Peng2021,
+   abstract = {This paper analyzes the factor zoo, which has theoretical and empirical implications for finance, from a machine learning perspective. More specifically, we discuss feature selection in the context of deep neural network models to predict the stock price direction. We investigated a set of 124 technical analysis indicators used as explanatory variables in the recent literature and specialized trading websites. We applied three feature selection methods to shrink the feature set aiming to eliminate redundant information from similar indicators. Using daily data from stocks of seven global market indexes between 2008 and 2019, we tested neural networks with different settings of hidden layers and dropout rates. We compared various classification metrics, taking into account profitability and transaction costs levels to analyze economic gains. The results show that the variables were not uniformly chosen by the feature selection algorithms and that the out-of-sample accuracy rate of the prediction converged to two values – besides the 50% accuracy value that would suggest market efficiency, a “strange attractor” of 65% accuracy also was achieved consistently. We also found that the profitability of the strategies did not manage to significantly outperform the Buy-and-Hold strategy, even showing fairly large negative values for some hyperparameter combinations.},
+   author = {Yaohao Peng and Pedro Henrique Melo Albuquerque and Herbert Kimura and Cayan Atreio Portela Bárcena Saavedra},
+   doi = {10.1016/j.mlwa.2021.100060},
+   issn = {26668270},
+   journal = {Machine Learning with Applications},
+   month = {9},
+   pages = {100060},
+   publisher = {Elsevier},
+   title = {Feature selection and deep neural networks for stock price direction forecasting using technical analysis indicators},
+   volume = {5},
+   url = {https://linkinghub.elsevier.com/retrieve/pii/S266682702100030X},
+   year = {2021},
+}
+@article{Fan2013,
+   abstract = {<p>The paper deals with the estimation of a high dimensional covariance with a conditional sparsity structure and fast diverging eigenvalues. By assuming a sparse error covariance matrix in an approximate factor model, we allow for the presence of some cross-sectional correlation even after taking out common but unobservable factors. We introduce the principal orthogonal complement thresholding method ‘POET’ to explore such an approximate factor structure with sparsity. The POET-estimator includes the sample covariance matrix, the factor-based covariance matrix, the thresholding estimator and the adaptive thresholding estimator as specific examples. We provide mathematical insights when the factor analysis is approximately the same as the principal component analysis for high dimensional data. The rates of convergence of the sparse residual covariance matrix and the conditional sparse covariance matrix are studied under various norms. It is shown that the effect of estimating the unknown factors vanishes as the dimensionality increases. The uniform rates of convergence for the unobserved factors and their factor loadings are derived. The asymptotic results are also verified by extensive simulation studies. Finally, a real data application on portfolio allocation is presented.</p>},
+   author = {Jianqing Fan and Yuan Liao and Martina Mincheva},
+   doi = {10.1111/rssb.12016},
+   issn = {1369-7412},
+   issue = {4},
+   journal = {Journal of the Royal Statistical Society Series B: Statistical Methodology},
+   keywords = {Approximate factor model,Cross-sectional correlation,Diverging eigenvalues,High dimensionality,Low rank matrix,Principal components,Sparse matrix,Thresholding,Unknown factors},
+   month = {9},
+   pages = {603-680},
+   pmid = {24348088},
[…]
