### 2024-08-19 a7c13c8
(paper) added drug interactions modeling application references per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2184909023
