+++ paper/paper.bib
@@ -208,3 +208,36 @@ programming mainstream.}
    url = {https://academic.oup.com/jrsssb/article/75/4/603/7075932},
    year = {2013},
 }
+
+
+@article{Guthrie2015,
+   abstract = {Background: The escalating use of prescribed drugs has increasingly raised concerns about polypharmacy. This study aims to examine changes in rates of polypharmacy and potentially serious drug-drug interactions in a stable geographical population between 1995 and 2010. Methods: This is a repeated cross-sectional analysis of community-dispensed prescribing data for all 310,000 adults resident in the Tayside region of Scotland in 1995 and 2010. The number of drug classes dispensed and the number of potentially serious drug-drug interactions (DDIs) in the previous 84 days were calculated, and age-sex standardised rates in 1995 and 2010 compared. Patient characteristics associated with receipt of ≥10 drugs and with the presence of one or more DDIs were examined using multilevel logistic regression to account for clustering of patients within primary care practices.},
+   author = {Bruce Guthrie and Boikanyo Makubate and Virginia Hernandez-Santiago and Tobias Dreischulte},
+   doi = {10.1186/s12916-015-0322-7},
+   issn = {1741-7015},
+   issue = {1},
+   journal = {BMC Medicine},
+   keywords = {Drug interactions,Family practice,Physician,Polypharmacy,Prescribing patterns,Primary care},
+   month = {12},
+   pages = {74},
+   title = {The rising tide of polypharmacy and drug-drug interactions: population database analysis 1995–2010},
+   volume = {13},
+   url = {http://bmcmedicine.biomedcentral.com/articles/10.1186/s12916-015-0322-7},
+   year = {2015},
+}
+@article{Maher2014,
+   abstract = {Introduction: Polypharmacy, defined as the use of multiple drugs or more than are medically necessary, is a growing concern for older adults. MEDLINE and EMBASE databases were searched from January 1, 1986 to June 30, 2013) to identify relevant articles in people aged > 65 years. Areas covered: We present information about: i) prevalence of polypharmacy and unnecessary medication use; ii) negative consequences of polypharmacy; and iii) interventions to improve polypharmacy. Expert opinion: International research shows that polypharmacy is common in older adults with the highest number of drugs taken by those residing in nursing homes. Nearly 50% of older adults take one or more medications that are not medically necessary. Research has clearly established a strong relationship between polypharmacy and negative clinical consequences. Moreover, well-designed interprofessional (often including clinical pharmacist) intervention studies that focus on enrolling high-risk older patients with polypharmacy have shown that they can be effective in reducing aspects of unnecessary prescribing with mixed results on distal health outcomes. © Informa UK, Ltd.},
+   author = {Robert L Maher and Joseph Hanlon and Emily R Hajjar},
+   doi = {10.1517/14740338.2013.827660},
+   issn = {1474-0338},
+   issue = {1},
+   journal = {Expert Opinion on Drug Safety},
+   keywords = {Aged,Drug utilization,Polypharmacy,Suboptimal drug use},
+   month = {1},
+   pages = {57-65},
+   publisher = {NIH Public Access},
+   title = {Clinical consequences of polypharmacy in elderly},
+   volume = {13},
+   url = {https://www.tandfonline.com/doi/full/10.1517/14740338.2013.827660},
+   year = {2014},
+}
