### 2024-08-19 592c3bc
(paper) added microbial interactions modeling application references per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2184909023
