+++ paper/paper.bib
@@ -241,3 +241,33 @@ programming mainstream.}
    url = {https://www.tandfonline.com/doi/full/10.1517/14740338.2013.827660},
    year = {2014},
 }
+
+
+@article{Leeuwen2023,
+   abstract = {<p>The human gut harbors native microbial communities, forming a highly complex ecosystem. Synthetic microbial communities (SynComs) of the human gut are an assembly of microorganisms isolated from human mucosa or fecal samples. In recent decades, the ever-expanding culturing capacity and affordable sequencing, together with advanced computational modeling, started a ‘‘golden age’’ for harnessing the beneficial potential of SynComs to fight gastrointestinal disorders, such as infections and chronic inflammatory bowel diseases. As simplified and completely defined microbiota, SynComs offer a promising reductionist approach to understanding the multispecies and multikingdom interactions in the microbe–host-immune axis. However, there are still many challenges to overcome before we can precisely construct SynComs of designed function and efficacy that allow the translation of scientific findings to patients’ treatments. Here, we discussed the strategies used to design, assemble, and test a SynCom, and address the significant challenges, which are of microbiological, engineering, and translational nature, that stand in the way of using SynComs as live bacterial therapeutics.</p>},
+   author = {Pim T van Leeuwen and Stanley Brul and Jianbo Zhang and Meike T Wortel},
+   doi = {10.1093/femsre/fuad012},
+   issn = {1574-6976},
+   issue = {2},
+   journal = {FEMS Microbiology Reviews},
+   keywords = {human gut microbiome,live bacterial therapeutic,rational design,synthetic microbial consortium},
+   month = {3},
+   title = {Synthetic microbial communities (SynComs) of the human gut: design, assembly, and applications},
+   volume = {47},
+   url = {https://academic.oup.com/femsre/article/doi/10.1093/femsre/fuad012/7080139},
+   year = {2023},
+}
+@article{Berg2022,
+   abstract = {M icrobial communities profoundly contribute to human and planetary health 1-5. Yet, the quantitative principles underlying community composition and assembly, and the link between community diversity and function, are largely unknown. Studying the emergent properties of microbial communities requires relating interactions across spatio-temporal scales, as well as their evolutionary dynamics. Unravelling, and ultimately predicting, community dynamics and emergent properties is a topical challenge in microbial ecology. Higher-order units-from microbial consortia to entire ecosystems feature macroscopic properties that emerge from microscopic interactions 6,7. Such emergent properties refer to any pattern or function that cannot be deduced as the sum of the properties of the constituent parts (Fig. 1). Examples in microbial systems include resilience to abiotic and biotic perturbations 8 , stable coexistence 9,10 and biochemical abilities 11-15. Emergent properties typically arise when community members reach some threshold level of community size and connectivity 16-18. Conditions such as high flow rates or high medium viscosity can obstruct connectivity and hence limit the onset of emergent properties 19,20. Finding the threshold con-nectivity is often difficult, even in a defined community, due to the non-linear nature of emergence. The threshold is property-specific and may also (non-linearly) depend on environmental factors 21. Hence, descriptions of emergent properties have been largely qualitative , although synthetic biology approaches are starting to characterize emergent properties quantitatively 14,22. Natural communities are difficult to study in situ due to their inherent dynamic nature and heterogeneity of habitats. Furthermore, abiotic variables influencing communities are difficult to measure or control in situ. Laboratory experiments, often considering the interactions between two or three community members, are amenable to quantitative analysis but limited in capturing emergent properties of natural, more diverse microbial communities. Defined communities with larger membership can be constituted in vitro 23 yet are limited to culturable microbes 24. Therefore, mathematical models are indispensable for linking community composition and connectivity to emergent functions. Models can bridge principles learned from simple laboratory systems to complex natural ecosystems ; this is a continui
[…]
