### 2024-08-23 amkrajewski/nimCSO#3 @RMeli
Hi @amkrajewski and @bdice is this issue still ongoing, or it can be resolved?
