### 2024-08-23 6bd376e
(paper) adjusted the algorithm-based search section to better explain the trade-offs between computational efficiency and solution optimality, clarify the algorithm's application context, and performance in edge cases per https://github.com/amkrajewski/nimCSO/issues/2#issuecomment-2295315929
