+++ paper/paper.md
@@ -86,8 +86,7 @@ The brute-force search is a naïve method of evaluating all possibilities; howev
 
 ## Algorithm-Based Search
 
-The algorithm implemented in the `algorithmSearch` routine, targeting high dimensional problems (20-50), iteratively expands and evaluates candidates from a priority queue (implemented through an efficient binary heap [@Williams1964]) while leveraging the fact that *the number of data points lost when removing elements `A` and `B` from the dataset has to be at least as large as when removing either `A` or `B` alone* to delay exploration of candidates until they can contribute to the solution. Furthermore, to (1) avoid revisiting the same candidate without keeping track of visited states and (2) further inhibit the exploration of unlikely candidates, the algorithm *assumes* that while searching for a given order of solution, elements present in already expanded solutions will not improve those not yet expanded. This effectively prunes candidate branches requiring two or more levels of backtracking. In the authors' tests, this method has generated the same results as `bruteForce`, except for occasional differences in the last explored solution.
-
+The algorithm implemented in the `algorithmSearch` routine, targeting high dimensional problems (20-50), iteratively expands and evaluates candidates from a priority queue (implemented through an efficient binary heap [@Williams1964]) while leveraging the fact that *the number of data points lost when removing elements `A` and `B` from the dataset has to be at least as large as when removing either `A` or `B` alone* to delay exploration of candidates until they *can* contribute to the solution. Furthermore, to (1) avoid revisiting the same candidate without keeping track of visited states and to (2) avoid exhaustive search by inhibiting the exploration of unlikely candidates, the algorithm *assumes* that while searching for a given order of solution, elements present in already expanded solutions will not improve those not yet expanded. This effectively prunes candidate branches requiring two or more levels of backtracking, making the algorithm computationally feasible in higher-dimensional problems at the cost of potentially suboptimal results. In the authors' tests over HEAs, this algorithm has generated correct results (agreeing with an exhaustive `bruteForce` exploration), except for occasional differences in the last explored solution, where almost all data was discarded, and a few remaining points were highly independent.
 
 ## Genetic Search
 
