### 2024-08-23 amkrajewski/nimCSO#3 @amkrajewski
I am still working on the last remaining point, i.e., adding a general-audience example of `nimCSO` use. I initially planned to use the suggested "pizza shop" example, which is excellent, but I decided to go with more general "ingredients in recipes" due to surprisingly extensive literature and data on the topic.

It sent me into a bit of a research rabbit hole, but I should be able to add it to the manuscript by this Monday. Additionally, after the revisions are complete, I plan to add a new tutorial using said ingredient data.
