### 2024-08-23 6848e24
(paper) added 2 references for the upcoming food recipes general-audience application discussion per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2185878853
