+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{Yeh2004,
-   abstract = {A new approach for the design of alloys is presented in this study. These "high-entropy alloys" with multi-principal elements were synthesized using well-developed processing technologies. Preliminary results demonstrate examples of 
+   abstract = {A new approach for the design of alloys is presented in this study. These "high-entropy alloys" with multi-principal elements were synthesized using well-developed processing technologies. Preliminary results demonstrate examples of
    the alloys with simple crystal structures, nanostructures, and promising mechanical properties. This approach may be opening a new era in materials science and engineering.},
    author = {Jien Wei Yeh and Swe Kai Chen and Su Jien Lin and Jon Yiew Gan and Tsung Shune Chin and Tao Tsung Shun and Chun Huei Tsau and Shou Yi Chang},
    doi = {10.1002/adem.200300567},
@@ -15,11 +15,11 @@
 
 
 @article{Cantor2004,
-   abstract = {Multicomponent alloys containing several components in equal atomic proportions have been manufactured by casting and melt spinning, and their microstructures and properties have been investigated by a combination of optical microscopy, 
-   scanning electron microscopy, electron probe microanalysis, X-ray diffractrometry and microhardness measurements. Alloys containing 16 and 20 components in equal proportions are multiphase, crystalline and brittle both as-cast and after melt 
-   spinning. A five component Fe20Cr20Mn20Ni20Co20 alloy forms a single fcc solid solution which solidifies dendritically. A wide range of other six to nine component late transition metal rich multicomponent alloys exhibit the same majority fcc primary 
-   dendritic phase, which can dissolve substantial amounts of other transition metals such as Nb, Ti and V. More electronegative elements such as Cu and Ge are less stable in the fcc dendrites and are rejected into the interdendritic regions. The total 
-   number of phases is always well below the maximum equilibrium number allowed by the Gibbs phase rule, and even further below the maximum number allowed under non-equilibrium solidification conditions. Glassy structures are not formed by casting or 
+   abstract = {Multicomponent alloys containing several components in equal atomic proportions have been manufactured by casting and melt spinning, and their microstructures and properties have been investigated by a combination of optical microscopy,
+   scanning electron microscopy, electron probe microanalysis, X-ray diffractrometry and microhardness measurements. Alloys containing 16 and 20 components in equal proportions are multiphase, crystalline and brittle both as-cast and after melt
+   spinning. A five component Fe20Cr20Mn20Ni20Co20 alloy forms a single fcc solid solution which solidifies dendritically. A wide range of other six to nine component late transition metal rich multicomponent alloys exhibit the same majority fcc primary
+   dendritic phase, which can dissolve substantial amounts of other transition metals such as Nb, Ti and V. More electronegative elements such as Cu and Ge are less stable in the fcc dendrites and are rejected into the interdendritic regions. The total
+   number of phases is always well below the maximum equilibrium number allowed by the Gibbs phase rule, and even further below the maximum number allowed under non-equilibrium solidification conditions. Glassy structures are not formed by casting or
    melt spinning of late transition metal rich multicomponent alloys, indicating that the confusion principle does not apply, and other factors are more important in promoting glass formation. © 2003 Elsevier B.V.},
    author = {B. Cantor and I. T.H. Chang and P. Knight and A. J.B. Vincent},
    doi = {10.1016/j.msea.2003.10.257},
@@ -35,18 +35,18 @@
 }
 
 @article{Williams1964,
[…]
