### 2024-08-26 amkrajewski/nimCSO#3 @RMeli
Thank you for the update! FYI, I'll be away for the next two weeks, apologies in advice if my replies will be slow.
