### 2024-08-26 amkrajewski/nimCSO#3 @bdice
@amkrajewski Sounds good to me. Ping me when you're finished making changes and I'll update this issue and my review checklist!
