### 2024-08-28 9d8ed0a
(paper) added food recipes general-audience application example to the Summary discussion per https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2185878853
