### 2024-08-29 7511dd7
(paper) mention that the Python and NumPy implementations are what is typically used in the community per https://github.com/amkrajewski/nimCSO/issues/2#issuecomment-2295315929
