### 2024-08-29 3b607c1
(quickstart) imporoved the tutorial introduction with general audience analogy to spicing food, which should clarify concepts of preventing data and removal of elements; also added analogies throughout the tutorial; per https://github.com/amkrajewski/nimCSO/issues/2#issuecomment-2295315929
