### 2024-08-29 amkrajewski/nimCSO#2 @amkrajewski
Hi @Henrium, thanks again for taking the time and effort to provide valuable feedback. I finished working through all reviewer comments, and I believe the code/paper is ready for you to evaluate my adjustments.

Notes:
- Above, please find the itemized list of changes I made (some overlapped with https://github.com/amkrajewski/nimCSO/issues/3).
- In some cases, I wanted to be more verbose, but the paper was already above the recommended length.
- I am happy to make further adjustments.
