### 2024-08-29 amkrajewski/nimCSO#3 @amkrajewski
Hi @bdice, thanks again for taking the time and effort to provide valuable feedback! I believe the code and paper adjustments are ready for you to evaluate.

On a side note, as mentioned in my last comment, I am working on an example showing `nimCSO` deployed over cooking ingredients and spices, but I think it can come after the journal submission is finalized.
