### 2024-09-04 amkrajewski/nimCSO#2 @Henrium
Hi @amkrajewski, thanks for your comprehensive response to my reviews. My concerns are well addressed, and I think the paper is in good shape @RMeli.
