### 2024-09-11 amkrajewski/nimCSO#2 @RMeli
Thanks @Henrium. @amkrajewski, I think we can close the issues then.

@Henrium, would you mind looking at your checklist in https://github.com/openjournals/joss-reviews/issues/6731#issuecomment-2099348790 and see if there are still outstanding items to be addressed? Many thanks!
