### 2024-09-11 amkrajewski/nimCSO#3 @RMeli
Hi @bdice, when you evaluate the latest changes from @amkrajewski, please update your checklist in https://github.com/openjournals/joss-reviews/issues/6731#issuecomment-2183177055 accordingly, so that we can keep track of the progress. Many thanks!
