### 2024-09-12 amkrajewski/nimCSO#2 @Henrium
Thanks for the reminder, all items have been addressed. I've updated the
checklist.
