### 2024-09-13 amkrajewski/nimCSO#2 @amkrajewski
@Henrium Again, thank you for your time and effort in the review. I will close this issue as completed.
