### 2024-09-14 #6731 @bdice
I am re-reviewing this and want the latest PDF. I think I need to run `@editorialbot generate pdf` so I'm going to do that.
