### 2024-09-14 amkrajewski/nimCSO#3 @bdice
@amkrajewski I checked most of the boxes above. I had one remaining questions from the first review that I didn't see addressed (please forgive me if I missed something!).

**References:** There's a link in the citations for Goldberg et al. that seems broken. Can you check it? https://doi.org/10.5860/choice.27-0936
