### 2024-09-14 #6731 @bdice
Status update: I had [one small question](https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2351108055) about a broken link in the references, but otherwise I recommend to accept! Thanks @amkrajewski for the revisions, I think the paper is a lot stronger now. 👍
