### 2024-09-16 41244d0
(paper) switched one of the book references from having poorly-resolving DOI to much more informative URL pointing to its page at ACM
