+++ paper/paper.bib
@@ -135,7 +135,7 @@ author = {Goldberg, David E.},
 title = {Genetic Algorithms in Search, Optimization and Machine Learning},
 year = {1989},
 isbn = {0201157675},
-doi = {10.5860/choice.27-0936},
+url = {https://dl.acm.org/doi/book/10.5555/534133},
 publisher = {Addison-Wesley Longman Publishing Co., Inc.},
 address = {USA},
 edition = {1st},
