### 2024-09-16 amkrajewski/nimCSO#3 @amkrajewski
@bdice Thanks for pointing that out! I verified the DOI was correct and led to the online version publisher (different than the physical, it seems). However, you are 100% right that a log-in paywall is not a great starting point for a reader looking for a reference. Thus, in https://github.com/amkrajewski/nimCSO/commit/2c70b196adc9dbba8584433a7da7ea8c40f25c40, I (1) removed the DOI so that the ISBN would show up as the primary reference, and (2) added a link to the ACM Digital Library page of the book which is quite helpful I think.
