### 2024-09-16 #6731 @bdice
@amkrajewski fixed up the last issue here: https://github.com/amkrajewski/nimCSO/issues/3#issuecomment-2351926925

I have no more blockers, this looks good to me.
