### 2024-09-18 amkrajewski/nimCSO#3 @amkrajewski
@bdice Thanks for providing an exceptional peer review! I will close the Issue now.
