### 2024-09-23 #6731 @RMeli
I left a few suggestions for the paper: https://github.com/amkrajewski/nimCSO/pull/6. DOIs look OK, the suggested missing one doesn't appear to be valid.
