### 2024-09-23 #6731 @RMeli
When you have reviewed and merged https://github.com/amkrajewski/nimCSO/pull/6 can you please make a new release of the software with all the changes from the review and provide the DOI for the archive? Thank you.
