### 2024-09-24 bcf4bf4
(paper) added `NumPy` citation per suggestion in #6
