### 2024-09-24 #6731 @amkrajewski
@RMeli The post-review version of code and paper has been released under version [**v0.6.5**](https://github.com/amkrajewski/nimCSO/releases/tag/v0.6.5) and published through Zenodo under doi:[**10.5281/zenodo.13834423**](https://zenodo.org/doi/10.5281/zenodo.13834423). I believe I followed all instructions, but I'm happy to make adjustments as needed.
