### 2024-10-17 7ba2d2a
- pointed main figure in `README` to permalink to display correctly outside GitHub (e.g., https://nimble.directory/pkg/nimcso)
