### 2024-10-28 #6731 @amkrajewski
Hi @RMeli ! It has been a couple of weeks, and the paper still hasn't gone through the publishing pipeline, so I was wondering if there is anything I should be doing on my part.

The [PR](https://github.com/openjournals/joss-papers/pull/5908) looks good; however, I saw that there was some error in the [associated GitHub Action](https://github.com/openjournals/joss-papers/actions/runs/11020647051/job/30605898294) related to JATS.

JATS is something I'm not too familiar with, but from what I gathered, it seems like compliance checks against a standard (within the publishing back-end) are possibly failing because my Table 1 has a double line on the bottom, making JATS expect content between them. Should I try adjusting that?
