### 2024-10-28 #6731 @RMeli
Hi @amkrajewski, thank you for the ping. I'm sorry but I missed the issue, and thought the paper was already handed over to the EiC.

I'm not very familiar either with this error, but looking at other submission it looks like it might be caused by your table formatting. Could you please try updating your table to use only official Markdown formatting? Thank you.
