### 2024-10-28 1bbea6e
(paper) reformatted the first table in order to increase its JATS compatibility within JOSS pipeline
