### 2024-11-14 #6731 @crvernon
🥳 Congratulations on your new publication @amkrajewski! Many thanks to @RMeli for editing and @atzberg, @Henrium, and @bdice for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
