### 2024-11-14 #6731 @amkrajewski
Once again, my coauthors and I would like to thank @bdice, @Henrium, and @atzberg for their excellent, constructive reviews that helped us further improve our tool and our Editor @RMeli for making our submission process a very positive experience! I'm eager to contribute to the JOSS community in the future and see it grow.
